AL-SHIFA-DENTAL-SYSTEM/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
email-validator
requests
chromadb
python-dotenv

# --- NEW AI AGENT LIBRARIES ---
google-generativeai
pypdf
pandas
apscheduler
openai
